Time: 11 ms (42.30%) | Memory: 8.3 MB (48.20%) - LeetSync

+++ 9-palindrome-number/palindrome-number.cpp
@@ -3,7 +3,7 @@ public:
     bool isPalindrome(int x) {
         
         long revNum = 0;
-        int dup = x;
+        long dup = x;
         
         while(x>0){
             int lastDigit = x%10;
